vsc-extension: move dependencyManager under mods directory
